zaakafhandelcomponent-1433 ACCEPTATIE BUG: Bij interne taak kan de inhoud van de radio buttons voor de adviseurs niet worden geconfigureerd
