+++ src/main/java/net/atos/zac/app/admin/converter/RESTReferentieTabelConverter.java
@@ -1,5 +1,5 @@
 /*
- * SPDX-FileCopyrightText: 2021 Atos
+ * SPDX-FileCopyrightText: 2022 Atos
  * SPDX-License-Identifier: EUPL-1.2+
  */
